% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_finder_function.R
\name{a_finder}
\alias{a_finder}
\title{a_finder}
\usage{
a_finder(qt, qc, b, g1)
}
\arguments{
\item{qt}{quantile 60 percent of future scenario}

\item{qc}{quantile 60 percent of controlled scenario}

\item{b}{parameter can be find by  b_finder function}

\item{g}{the difference between 60 percent quantiles of observed and control scenario, "g1"}
}
\value{
a numeric value
}
\description{
a_finder
}
