% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_finder_function.R
\name{b_finder}
\alias{b_finder}
\title{b_finder}
\usage{
b_finder(d, l, g, h, c, o)
}
\arguments{
\item{d}{the difference between 90 percent quantiles "g2"}

\item{l}{quantile 90 percent of future scenario}

\item{g}{the difference between 60 percent quantiles "g1"}

\item{h}{quantile 60 percent of future scenario}

\item{c}{quantile 90 percent of controlled time}

\item{o}{quantile 60 percent of controlled time}
}
\value{
numeric value
}
\description{
b_finder
}
