% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plots_function.R
\name{compare_plots}
\alias{compare_plots}
\title{comparer_plots}
\usage{
compare_plots(ob_data, sim_data)
}
\arguments{
\item{ob_data}{your observed data in form of data frame}

\item{sim_data}{your simulated data in a form of data frame}
}
\value{

}
\description{
this function plot two different data to compare them with each other , it mostly used to compare observed data and simulated data
}
