% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_log10_prior}
\alias{compute_log10_prior}
\title{Compute log10 prior of model configuration}
\usage{
compute_log10_prior(mcfg, pi_vec)
}
\arguments{
\item{mcfg}{A vector of model configuration}

\item{pi_vec}{A vector of prior probabilities}
}
\value{
Result Log 10 prior of model configuration
}
\description{
Compute log10 prior of model configuration
}
