% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_inf.R
\name{cred}
\alias{cred}
\title{Compute credible sets from single-effect PIPs}
\usage{
cred(
  PIP,
  coverage = 0.9,
  purity = 0.5,
  LD = NULL,
  V = NULL,
  Dsq = NULL,
  n = NULL,
  dedup = TRUE
)
}
\arguments{
\item{PIP}{p x L PIP matrix output by susie}

\item{coverage}{Coverage level for each credible set}

\item{purity}{Sets with minimum absolute correlation < purity are removed}

\item{LD}{LD matrix (equal to X'X/n)}

\item{V}{Precomputed p x p matrix of eigenvectors of X'X}

\item{Dsq}{Precomputed length-p vector of eigenvalues of X'X}

\item{n}{Sample size}

\item{dedup}{Remove duplicate CS's}
}
\value{
List of variable indices corresponding to credible sets
}
\description{
Compute credible sets from single-effect PIPs
}
