% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagedocumentation.R
\docType{package}
\name{dap-package}
\alias{dap-package}
\title{Discriminant Analysis via Projections}
\description{
This package provides tools for discriminant analysis on binary classification. It contains functions \code{apply_DAP, classify_DAP, cv_DAP, solve_DAP_C, solve_DAP, solve_DAP_seq} for implementing the method Discriminant Analysis via Projections.
}
\references{
Gaynanova, I. and Wang, T. "Sparse quadratic classification rules via linear dimension reduction". arxiv.org/abs/1711.04817 (2018+)
}
\author{
Irina Gaynanova and Tianying Wang.
}
\keyword{package}
