% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dap}
\alias{dap}
\title{Integrative Genetic Association Analysis using Deterministic Approximation of Posteriors}
\usage{
dap(file="/path/to/sbams/file")
dap(sbams=a.sbams.object)
dap(ld="/path/to/ld/file", est="/path/to/est/file", n, syy)
}
\arguments{
\item{file}{file path to a sbams file}

\item{sbams}{a sbams object}

\item{ld}{file path to a ld file}

\item{est}{file path to a est file}

\item{n}{an integer indicating the sample size}

\item{syy}{a float number indicating the total sum of squares for the quantitative trait}
}
\value{
a dap object including model summary, SNP summary and signal cluster summary
}
\description{
The \code{dap} function accepts two different sets of inputs.
\enumerate{
\item sbams (either file path or object)
\item summary statistics (ld, est, n, syy)
}
}
\details{
The summary statistics can be obtained by \code{summary(a.sbams.object)}.
}
