% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dap.ss}
\alias{dap.ss}
\title{dap.ss}
\usage{
dap.ss(est, ld, n, syy)
}
\description{
dap.ss
}
\examples{
\dontrun{

est_file = system.file("sbamsdat", "sim.1.est.dat", package = "dap")
ld_file  = system.file("sbamsdat", "sim.1.ld.dat", package = "dap")
est = read.table(est_file)
ld = read.table(ld_file)
ld = as.matrix(ld)
dap.ss(est, ld, 343, 515.6)
}
}
