% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap_inf.R
\name{dap_inf}
\alias{dap_inf}
\title{DAP-S fine mapping with infinitesimal effects}
\usage{
dap_inf(
  bhat = NULL,
  shat = NULL,
  z = NULL,
  var_y = NULL,
  n,
  LD = NULL,
  L = 10,
  prior_weights = NULL,
  standardize = TRUE,
  use_susie_variance_estimate = TRUE,
  grid = c(0.04, 0.16, 0.64),
  overlapping = TRUE,
  pir_threshold = 1e-06,
  r2_threshold = 0.25,
  coverage = NULL,
  method = "moments",
  twas_weight = FALSE
)
}
\arguments{
\item{bhat}{Estimated effect size}

\item{shat}{Standard error of bhat}

\item{z}{Vector of z-scores}

\item{var_y}{Variance of y}

\item{n}{Sample size}

\item{LD}{LD matrix}

\item{L}{Number of modeled causal effects}

\item{prior_weights}{Vector of prior probabilities for each variant being
causal. Default is uniform priors}

\item{standardize}{Standardize each column of the genotype matrix. Default is
TRUE.}

\item{use_susie_variance_estimate}{Use SuSiE-estimated prior variance to
guide the grid upper bound. Default is TRUE.}

\item{grid}{Grid of prior variances, default is c(0.04, 0.16, 0.64). The
upper bound is shifted to 1.28 if the maximum of SuSiE-estimated prior
variance is greater than 1.}

\item{overlapping}{Overlapping signal clusters. Default is TRUE.}

\item{pir_threshold}{Threshold for PIR. Default is 1e-6.}

\item{r2_threshold}{Threshold for genotype R2 when constructing signal
clusters. Default is 0.25.}

\item{coverage}{Coverage for credible set. If NULL, the algorithm will
construct signal clusters.}

\item{twas_weight}{Compute TWAS weights. Default is FALSE.}

\item{null_weight}{Null weight used in SuSiE. Default is calculated as
\code{prod(1 - prior_weights)}.}
}
\value{
List of fine-mapping results.
}
\description{
DAP-S fine mapping with infinitesimal effects
}
\examples{
# Set random seed for reproducibility
set.seed(2025)
n <- 5000  # Number of samples
p <- 500   # Number of SNPs
MAF <- 0.1  # Minor allele frequency
Ltrue <- 5  # Number of true causal variants
ssq <- 0.01  # Effect size variance
sigmasq <- 1  # Error variance

# Generate genotype matrix X
X <- matrix(rbinom(n*p, 2, MAF), nrow=n, ncol=p)
X <- scale(X, center=TRUE, scale=FALSE)

# Generate sparse effects
b <- rep(0, p)
inds <- sample(1:p, size=Ltrue, replace=FALSE)
b[inds] <- rnorm(Ltrue) * sqrt(ssq)
order_idx <- order(inds)
cat('True effects:', inds[order_idx], '\n')
cat('Effect sizes:', b[inds[order_idx]], '\n')

# Generate infinitesimal effects
tausq <- 1e-3
infinitesimal <- X \%*\% rnorm(p) * sqrt(tausq)
effects <- X \%*\% b + infinitesimal
y <- effects + rnorm(n) * sqrt(sigmasq)
y <- scale(y, center = TRUE, scale = FALSE)
cat('Total fraction of variance explained by SNPs:', var(as.vector(effects))/var(y), '\n')

# Generate summary statistics
res = susieR::univariate_regression(X, y)
suff_stat <- susieR::compute_suff_stat(X, as.vector(y), standardize = FALSE)
LD <- cov2cor(suff_stat$XtX)
z <- res$betahat/res$sebetahat

# Work on original scale
rst <- dap_inf(bhat = res$betahat, shat = res$sebetahat, var_y = var(y), n=n, L=5, LD = LD, twas_weight=TRUE)
rst$sigmasq
rst$tausq

# SuSiE-inf
output <- susie_inf(bhat = res$betahat, shat = res$sebetahat, var_y = var(y), n=n, L=5, LD = LD, null_weight = (1-1/p)^p)

# PIP comparison
plot(output$spip, rst$pip, xlab = "PIP of SuSiE-inf", ylab = "PIP of DAP-S")
abline(a = 0,b = 1,col = "skyblue",lty = "dashed")

# TWAS weight
plot(rowSums(output$mu * output$PIP)[1:p], rst$twas_weights,
    xlab = "Coefficients (sparse) of SuSiE-inf",
    ylab = "Coefficients (sparse) of DAP-S") + abline(a=0,b=1)
plot(rowSums(output$mu * output$PIP)[1:p] + output$alpha[1:p], rst$twas_weights,
    xlab = "Coefficients (sparse+infinitesimal) of SuSiE-inf",
    ylab = "Coefficients (sparse) of DAP-S") + abline(a=0,b=1)
}
