% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dap_main}
\alias{dap_main}
\title{Main function for DAP-PIR algorithm}
\usage{
dap_main(X, y, L, matrix, threshold, prior_weights, phi2_vec)
}
\arguments{
\item{X}{Genotype matrix}

\item{y}{Phenotype vector}

\item{L}{Number of causal variants}

\item{matrix}{Proposal density matrix from SuSiE}

\item{threshold}{Threshold for proposal density}

\item{prior_weights}{Vector of prior probabilities}

\item{phi2_vec}{Vector of scaled prior effect size variances}
}
\value{
A list containing:
\itemize{
  \item model_config - Model configurations
  \item posterior_prob - Posterior probabilities
  \item log10_posterior_score - Log10 posterior scores
  \item log10_nc - Log10 normalizing constant
  \item pip - Posterior inclusion probabilities
}
}
\description{
Performs fine mapping using DAP-PIR algorithm
}
