% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap_ss.R
\name{dap_rss}
\alias{dap_rss}
\title{DAP-S fine-mapping using summary statistics data}
\usage{
dap_rss(
  z,
  R,
  n,
  bhat,
  shat,
  var_y,
  estimate_residual_variance = FALSE,
  prior_variance = 50
)
}
\arguments{
\item{z}{The z-score matrix}

\item{R}{The correlation matrix}

\item{n}{The sample size}

\item{bhat}{The effect size estimates}

\item{var_y}{The variance of the phenotype}

\item{estimate_residual_variance}{Whether to estimate the residual variance}
}
\value{
The residual sum of squares
}
\description{
DAP-S fine-mapping using summary statistics data
}
