% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap_ss.R
\name{dap_suff_stat}
\alias{dap_suff_stat}
\title{DAP-S fine-mapping using sufficient statistics data}
\usage{
dap_suff_stat(
  XtX,
  Xty,
  yty,
  n,
  L = min(10, ncol(XtX)),
  susie_params = list(),
  prior_weights = NULL,
  standardize = TRUE,
  use_susie_variance_estimate = TRUE,
  grid = c(0.04, 0.16, 0.64),
  overlapping = TRUE,
  pir_threshold = 1e-06,
  r2_threshold = 0.25,
  coverage = NULL,
  twas_weight = TRUE
)
}
\arguments{
\item{XtX}{A p by p matrix \eqn{X'X} in which the columns of X are centered to have mean zero.}

\item{Xty}{A p-vector \eqn{X'y} in which y and the columns of X are centered to have mean zero.}

\item{yty}{A scalar \eqn{y'y} in which y is centered to have mean zero.}

\item{n}{The sample size.}

\item{L}{Number of causal variants}

\item{prior_weights}{Prior weights}

\item{r2_threshold}{Genotype R2 threshold for LD}

\item{coverage}{Coverage of credible sets. When not set, it outputs signal clusters;
otherwise, it outputs credible sets at the specified coverage level}

\item{null_weight}{Null weight}

\item{residual_tau}{Residual variance}

\item{threshold}{Threshold for proposal density}

\item{phi2}{Scaled prior effect size variance}

\item{phi2_vec}{Scaled prior effect size variance vector}
}
\value{
Fine mapping results
}
\description{
DAP-S fine-mapping using sufficient statistics data
}
