% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap_susie.R
\name{dap_susie}
\alias{dap_susie}
\title{DAP-S fine-mapping with individual-level data. The parameters are directly obtained from SuSiE results. # nolint: line_length_linter.}
\usage{
dap_susie(
  X,
  y,
  L = min(10, ncol(X)),
  scaled_prior_variance = 0.2,
  residual_variance = NULL,
  prior_weights = NULL,
  null_weight = NULL,
  estimate_residual_variance = TRUE,
  estimate_prior_variance = TRUE,
  r2_threshold = 0.25,
  coverage = NULL,
  use_susie_variance_estimate = FALSE,
  phi2_vec = c(0.04, 0.16, 0.64),
  greedy = FALSE,
  pir_threshold = 1e-06
)
}
\arguments{
\item{X}{Genotype data}

\item{y}{Phenotype data}

\item{L}{Number of causal variants}

\item{scaled_prior_variance}{Scaled prior effect size variance. The prior variance, divided by \code{var(y)}; that is, the prior variance of each non-zero element of b is \code{var(y) * scaled_prior_variance}. So it is equal to phi^2. # nolint: line_length_linter.}

\item{residual_variance}{Residual variance. If \code{estimate_residual_variance = TRUE}, this value provides the initial estimate of the residual variance. # nolint: line_length_linter.}

\item{prior_weights}{Prior weights}

\item{null_weight}{Null weight. Probability of no effects, it should be between 0 and 1. # nolint: line_length_linter.}

\item{estimate_residual_variance}{If \code{estimate_residual_variance = TRUE}, the residual variance is estimated, using \code{residual_variance} as an initial value. If \code{estimate_residual_variance = FALSE}, the residual variance is fixed to the value supplied by \code{residual_variance}. # nolint: line_length_linter.}

\item{estimate_prior_variance}{If \code{estimate_prior_variance = TRUE}, the prior variance is estimated (this is a separate parameter for each of the L effects). If provided, \code{scaled_prior_variance} is then used as an initial value for the optimization. When \code{estimate_prior_variance = FALSE}, the prior variance for each of the L effects is determined by the value supplied to \code{scaled_prior_variance}. # nolint: line_length_linter.}

\item{r2_threshold}{Genotype R2 threshold for LD}

\item{coverage}{Coverage of credible sets. When not set, it outputs signal clusters; otherwise, it outputs credible sets at the specified coverage level # nolint: line_length_linter.}

\item{use_susie_variance_estimate}{If \code{use_susie_variance_estimate = TRUE}, the SuSiE variance estimate (V) is used to compute the Bayes factor. If \code{use_susie_variance_estimate = FALSE}, the Bayes factor is calculated by averaging over a grid of values provided in \code{phi2_vec}. # nolint: line_length_linter.}

\item{phi2_vec}{Scaled prior effect size variance vector. If \code{use_susie_variance_estimate = FALSE}, this vector is used to compute the Bayes factor. # nolint: line_length_linter.}

\item{greedy}{If \code{greedy = TRUE}, the greedy algorithm is used to find the best model, instead of conducting Pseudo-Importance Resampling. # nolint: line_length_linter.}

\item{pir_threshold}{Threshold for PIR when \code{greedy = FALSE}. # nolint: line_length_linter.}
}
\value{
Fine mapping results
}
\description{
DAP-S fine-mapping with individual-level data. The parameters are directly obtained from SuSiE results. # nolint: line_length_linter.
}
