% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dap_susie_main}
\alias{dap_susie_main}
\title{Implementation of DAP-S algorithm in C++ with default SuSiE settings}
\usage{
dap_susie_main(
  X,
  y,
  matrix,
  prior_weights,
  r2_threshold,
  coverage,
  use_susie_variance_estimate,
  phi2_vec,
  greedy,
  pir_threshold
)
}
\arguments{
\item{X}{Genotype matrix}

\item{y}{Phenotype vector}

\item{matrix}{Proposal density matrix from SuSiE}

\item{prior_weights}{Vector of prior probabilities}

\item{r2_threshold}{Threshold for LD}

\item{coverage}{Coverage for credible set}

\item{use_susie_variance_estimate}{Use SuSiE variance estimate}

\item{phi2_vec}{Vector of scaled prior effect size variances}

\item{greedy}{Use greedy algorithm}

\item{pir_threshold}{Threshold for PIR}
}
\value{
A list containing:
\itemize{
  \item model_config - Model configurations
  \item posterior_prob - Posterior probabilities
  \item log10_posterior_score - Log10 posterior scores
  \item log10_nc - Log10 normalizing constant
  \item pip - Posterior inclusion probabilities
}
}
\description{
Performs fine mapping using DAP-S algorithm
}
