% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap_update.R
\name{dap_update}
\alias{dap_update}
\title{Update DAP-S fine-mapping results.}
\usage{
dap_update(
  X,
  dap_result,
  prior_weights = NULL,
  r2_threshold = 0.25,
  coverage = NULL
)
}
\arguments{
\item{X}{Genotype matrix (n x p)}

\item{dap_result}{DAP-S result from \code{dap}}

\item{prior_weights}{New prior weights of each SNP being causal}

\item{r2_threshold}{Threshold for R2 when constructing credible sets}

\item{coverage}{Coverage for credible set}
}
\value{
List of fine-mapping results.
}
\description{
Update DAP-S fine-mapping results.
}
