% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_finder_function.R
\name{dif_finder}
\alias{dif_finder}
\title{difference finder}
\usage{
dif_finder(a, b)
}
\arguments{
\item{a}{numerical input}

\item{b}{numerical input}
}
\value{
numerical value
}
\description{
difference finder
}
\examples{
dif_finder(4,2) = 2
}
