% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbams.R
\name{extract.sbams}
\alias{extract.sbams}
\title{Extract Sufficient Statistics from a sbams file}
\usage{
extract.sbams(file, ens = 1, pi1 = -1)
}
\arguments{
\item{file}{file path to the sbams file}

\item{ens}{(optional) prior expected number of signals, \code{ens=1} by default.}

\item{pi1}{(optional) the exchangeable prior probability, values \code{0<pi1<1} accepted. By default -1, \code{pi1=ens/p}, where \code{p} is number of SNPs in the input file.}
}
\value{
\code{extract.sbams} returns a list containing the following components: \item{est}{a \code{data.frame} including snp names, estimated effect size and corresponding estimated standard error.}
\item{ld}{a \code{matrix} representing the correlation matrix of SNPs.}
\item{n}{sample size.}
\item{syy}{total sum of squares for the quantitative phenotype.}
\item{name}{name of the phenotype}
}
\description{
The sufficient summary statistics refer to the following information: \itemize{
\item estimated effect size and corresponding estimated standard error from single SNP testing for each SNP
\item correlation matrix of SNPs
\item sample size
\item total sum of squares for the quantitative phenotype
}
}
\examples{
sbams.file = system.file("sbamsdat", "sim.1.sbams.dat", package = "dap")
ss = extract.sbams(sbams.file)
dap.ss(ss$est, ss$ld, ss$n, ss$syy, ss$name)

}
