% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracter_function.R
\name{extract_my_data}
\alias{extract_my_data}
\title{extractor}
\usage{
extract_my_data(from, by)
}
\arguments{
\item{from}{input "raster" that you want to extract from}

\item{by}{input "vector" that you want to extract by}
}
\value{
a raster
}
\description{
extract chosen values from a raster
}
