% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_exceedance_finder_function.R
\name{mean_ex_finder}
\alias{mean_ex_finder}
\title{mean_ex_finder}
\usage{
mean_ex_finder(pr, q)
}
\arguments{
\item{pr}{precipitation}

\item{q}{quantile 90 percent}
}
\value{
a numeric value
}
\description{
this function finds the mean exeedance ,the value choosen for precipitation better be greater than quantile 90 percent
}
