% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_inf.R
\name{MLE}
\alias{MLE}
\title{Maximum Likelihood Estimation (MLE) for sigma^2 and tau^2}
\usage{
MLE(
  PIP,
  mu,
  omega,
  sigmasq,
  tausq,
  n,
  V,
  Dsq,
  VtXty,
  yty,
  est_sigmasq,
  est_tausq,
  sigmasq_range,
  tausq_range,
  it,
  verbose
)
}
\arguments{
\item{PIP}{Matrix of posterior inclusion probabilities}

\item{mu}{Matrix of posterior means conditional on causal}

\item{omega}{Matrix of posterior precisions conditional on causal}

\item{sigmasq}{Current estimate of sigma^2}

\item{tausq}{Current estimate of tau^2}

\item{n}{Sample size}

\item{V}{Matrix of eigenvectors of X'X}

\item{Dsq}{Eigenvalues of X'X}

\item{VtXty}{V'X'y}

\item{yty}{y'y}

\item{est_sigmasq}{Whether to estimate sigma^2}

\item{est_tausq}{Whether to estimate tau^2}

\item{sigmasq_range}{Range for sigma^2 optimization}

\item{tausq_range}{Range for tau^2 optimization}

\item{it}{Current iteration}

\item{verbose}{Whether to print update information}
}
\value{
List containing updated sigmasq and tausq
}
\description{
Maximum Likelihood Estimation (MLE) for sigma^2 and tau^2
}
