% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data_documentation.R
\docType{data}
\name{model_data}
\alias{model_data}
\title{model data}
\format{
a data table with 3 columns and 55115 rows  that includes:
\describe{
  \item{DTM}{time of recording data}
  \item{pr}{precipitation value}
  \item{tem}{tempreture value }

}
}
\usage{
model_data
}
\description{
showing the precipitation and tempreture data of simulated model
}
\keyword{datasets}
