% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_inf.R
\name{MoM}
\alias{MoM}
\title{Method of Moments (MoM) estimation for sigma^2 and tau^2}
\usage{
MoM(
  PIP,
  mu,
  omega,
  sigmasq,
  tausq,
  n,
  V,
  Dsq,
  VtXty,
  Xty,
  yty,
  est_sigmasq,
  est_tausq,
  verbose
)
}
\arguments{
\item{PIP}{Matrix of posterior inclusion probabilities}

\item{mu}{Matrix of posterior means conditional on causal}

\item{omega}{Matrix of posterior precisions conditional on causal}

\item{sigmasq}{Current estimate of sigma^2}

\item{tausq}{Current estimate of tau^2}

\item{n}{Sample size}

\item{V}{Matrix of eigenvectors of X'X}

\item{Dsq}{Eigenvalues of X'X}

\item{VtXty}{V'X'y}

\item{Xty}{X'y}

\item{yty}{y'y}

\item{est_sigmasq}{Whether to estimate sigma^2}

\item{est_tausq}{Whether to estimate tau^2}

\item{verbose}{Whether to print update information}
}
\value{
List containing updated sigmasq and tausq
}
\description{
Method of Moments (MoM) estimation for sigma^2 and tau^2
}
