% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_data_documentation.R
\docType{data}
\name{observed_data}
\alias{observed_data}
\title{observed Klementinum data}
\format{
a data table with 3 columns and 73322 rows that includes:
\describe{
  \item{DTM}{recording time}
  \item{pr_o}{observed precipitation}
  \item{tem_o}{observed tempreture}
}
}
\usage{
observed_data
}
\description{
the data include observed precipitation and temperature for Klementinum
}
\keyword{datasets}
