% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pir}
\alias{pir}
\title{Find all model combinations that have a proposal density greater than or equal to a threshold}
\usage{
pir(mat, threshold)
}
\arguments{
\item{mat}{A processed p x l matrix of proposal densities from SuSiE}

\item{threshold}{The threshold value of the proposal density}
}
\value{
A NumericMatrix containing the unique combinations
}
\description{
Find all model combinations that have a proposal density greater than or equal to a threshold
}
