% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_raster_documentation.R
\docType{data}
\name{pr_raster}
\alias{pr_raster}
\title{simulated  data}
\format{
a raster with dimension  16, 17, 272  (nrow, ncol, ncell) that includes:
 \describe{
  \item{z-value}{recorded time}
  \item{zvar}{simulated precipitation}
  \item{band}{1  (of  55115  bands)}
  \item{resolution}{2.8125, 2.790426  (x, y)}
  \item{extent}{-1.40625, 46.40625, 27.90615, 72.55296  (xmin, xmax, ymin, ymax)}
  \item{crs}{+proj=longlat +datum=WGS84}
}
}
\usage{
pr_raster
}
\description{
the data include simulated precipitation
}
\keyword{datasets}
