% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_functions.R
\name{pr_trans_under}
\alias{pr_trans_under}
\title{pr_trans_under}
\usage{
pr_trans_under(a, pr, b)
}
\arguments{
\item{a}{parameter ,can be find by a_finder function}

\item{pr}{observed precipitation}

\item{b}{parameter , can be find by b_finder function}
}
\value{
numeric value
}
\description{
transition of precipitation under 90 percent quantile.the value of observed precipitation given to the function should be less than 90 percent quentile
}
