% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_functions.R
\name{pr_trans_upon}
\alias{pr_trans_upon}
\title{pr_trans_upon}
\usage{
pr_trans_upon(E, pr, q, a, b)
}
\arguments{
\item{E}{the difference between mean exceedance of future and control scenario}

\item{pr}{observed precipitation ,the value of observed precipitation given to the function should be greater than 90 percent quentile}

\item{q}{quantile 90 percent}

\item{a}{parameter, you can find it by a_finder function}

\item{b}{parameter, you can find it by b_finder function}
}
\value{

}
\description{
transition of precipitation upon 90 percent quantile.the value of observed precipitation given to the function should be greater than 90 percent quentile
}
