% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{scan}
\alias{scan}
\title{Single-SNP Association Analysis}
\usage{
scan(formula, data)
}
\arguments{
\item{formula}{an object of class \code{formula} (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{a data frame containing the variables in the model.}
}
\value{
\code{scan} returns a \code{data.frame} containing effect estimates for each single SNP.
}
\description{
\code{scan} is used perform perform single-SNP fine-mapping analysis.
}
\examples{
set.seed(0)
n = 100
p = 1000

a1 = rnorm(n)
a2 = 10*a1+rnorm(n)
a3 = 2*a1+9*a2+rnorm(n)

b1 = rnorm(n)
b2 = 8*b1+rnorm(n)

x = matrix(rnorm(n*(p-5)), nrow=n)

df = data.frame(a1,a2,a3,b1,b2,x)
df$y = 2*a1+b1+rnorm(n)

test.scan = scan(y~., df)
head(test.scan)

}
\seealso{
\code{\link{scan.sbams}} for a different interface directly analyzing a sbams-format file.
}
