% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{scan.sbams}
\alias{scan.sbams}
\title{Single-SNP Association Analysis for SBAMS-format Data}
\usage{
scan.sbams(file)
}
\arguments{
\item{file}{file path to a sbams file}
}
\value{
\code{scan.sbams} returns a \code{data.frame} containing effect estimates for each single SNP.
}
\description{
\code{scan.sbams} is an interface built especially for SBAMS-format files, which is designed to perform single-SNP fine-mapping analysis.
}
\examples{
\dontrun{

sbams.file = system.file("sbamsdat", "sim.1.sbams.dat", package = "dap")
test.scan.sbams = scan.sbams(sbams.file)
head(test.scan.sbams)

}
}
\seealso{
\code{\link{read.sbams}} for reading in sbams-format files as an \R \code{data.frame} which can call the general version of \code{\link{scan}} function.
}
