% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_on_map.R
\name{show_on_map}
\alias{show_on_map}
\title{show_on_map}
\usage{
show_on_map(place, lon, lat)
}
\arguments{
\item{place}{a data frame including the place with latitude and longitude}

\item{lon}{column name of your data frame that includes longitude values}

\item{lat}{column name of your data frame that includes latitude values}
}
\value{
a map
}
\description{
this function shows your places as points on map
}
