% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statics_documentation.R
\docType{data}
\name{statics}
\alias{statics}
\title{statics}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 52 rows and 14 columns.
}
\usage{
statics
}
\description{
shows info such as quantile and mean exceedance
}
\details{
\itemize{
  \item block.  each 7 days became a block
  \item q60_o.  quantile 60 percent of observation data
  \item q90_o.  quantile 90 percent of observation data
  \item q60_c.  quantile  60 percent of controlled data
  \item q90_c.  quantile 90 percent of controlled data
  \item q60_f.   quantile 60 percent of future data
  \item q90_f.   quantile 90 percent of future data
  \item e_c.   mean exceedance of controlled data
  \item e_f.   mean exceedance of future data
  \item mean_tem_o.   mean observation temprature of each  block
  \item mean_tem_cont.   mean controlled temprature of each block
  \item mean_tem_futr.   mean future temprature of each block
  \item sdt_c.   standard diviation of temprature for controll time for each block
  \item sdt_f.   standard diviation of temprature for future time for each block

}
}
\keyword{datasets}
