% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_documentation.R
\docType{data}
\name{station}
\alias{station}
\title{Klementinum station coordinates}
\format{
a data table with 2 columns and 1 row  that includes:
\describe{
  \item{lon}{longitude}
  \item{lat}{latitude}

}
}
\usage{
station
}
\description{
showing the coordinates of Klementinum CZECH REPUBLIC,+50:05:11,+014:24:59,191 m
}
\keyword{datasets}
