% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{summary.dap}
\alias{summary.dap}
\title{Summarizing DAP results}
\usage{
\method{summary}{dap}(object)

\method{print}{summary.dap}(object, digits = max(5L, getOption("digits") - 3L))
}
\arguments{
\item{object}{an object of class \code{"dap"}, usually, a result of a call to \code{\link{dap}}.}

\item{digits}{the number of significant digits to use when printing.}
}
\value{
The function \code{summary.dap} returns a summary of dap results given in \code{object}, using its components \code{"call"} and \code{"info"}, plus \item{top.models}{a data frame for the top 5 (or less if not available) models with highest posterior probabilities, including model configuration, number of predictors involved, the posterior probability of the corresponding model, and the unnormalized posterior score of the model, defined as \eqn{log10(model prior)+log10(BF)}.}
\item{clusters}{a data frame for the signal cluster summary, including the size of the cluster (i.e. the number of member predictors), the corresponding posterior inclusion probability, and the remaining columns represent the average LD measures (\eqn{r^2}) for predictors within a cluster and between clusters.}
\item{signals}{a list of data frames for predictors in each cluster, including the predictor name, and corresponding posterior inclusion probability.}
}
\description{
\code{summary} method for class \code{"dap"}.
}
