% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_inf.R
\name{susie_inf}
\alias{susie_inf}
\title{SuSiE with random effects}
\usage{
susie_inf(
  z,
  bhat,
  shat,
  var_y,
  n,
  L,
  LD = NULL,
  V = NULL,
  Dsq = NULL,
  est_ssq = TRUE,
  ssq = NULL,
  ssq_range = c(0, 1),
  prior_weights = NULL,
  null_weight = 0,
  est_sigmasq = TRUE,
  est_tausq = TRUE,
  sigmasq = 1,
  tausq = 0,
  method = "moments",
  sigmasq_range = NULL,
  tausq_range = NULL,
  PIP = NULL,
  mu = NULL,
  maxiter = 100,
  PIP_tol = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{z}{Vector of z-scores (equal to X'y/sqrt(n))}

\item{bhat}{Beta hat}

\item{shat}{Standard error hat}

\item{var_y}{Variance of y, equal to ||y||^2/(n-1)}

\item{n}{Sample size}

\item{L}{Number of modeled causal effects}

\item{LD}{LD matrix (equal to X'X/n)}

\item{V}{Precomputed p x p matrix of eigenvectors of X'X}

\item{Dsq}{Precomputed length-p vector of eigenvalues of X'X}

\item{est_ssq}{Estimate prior effect size variances s^2 using MLE}

\item{ssq}{Length-L initialization s^2 for each effect}

\item{ssq_range}{Lower and upper bounds for each s^2, if estimated}

\item{prior_weights}{Length-p vector of prior causal probability for each SNP; must sum to 1}

\item{null_weight}{A value specify the probability of no effect; should be between 0 and 1.}

\item{est_sigmasq}{Estimate variance sigma^2}

\item{est_tausq}{Estimate both variances sigma^2 and tau^2}

\item{sigmasq}{Initial value for sigma^2}

\item{tausq}{Initial value for tau^2}

\item{method}{One of c('moments','MLE')}

\item{sigmasq_range}{Lower and upper bounds for sigma^2, if estimated using MLE}

\item{tausq_range}{Lower and upper bounds for tau^2, if estimated using MLE}

\item{PIP}{p x L initializations of PIPs}

\item{mu}{p x L initializations of mu}

\item{maxiter}{Maximum number of SuSiE iterations}

\item{PIP_tol}{Convergence threshold for PIP difference between iterations}

\item{verbose}{Whether to print update information}
}
\value{
List with results
}
\description{
SuSiE with random effects
}
