% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templot_function.R
\name{tem_plot}
\alias{tem_plot}
\title{tem_plot}
\usage{
tem_plot(ms, mc)
}
\arguments{
\item{ms}{mean tempreture of scenario period}

\item{mc}{mean tempreture of control period}
}
\value{
a plot
}
\description{
show the changes of tempreture
}
