% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_functions.R
\name{tem_trans_finder}
\alias{tem_trans_finder}
\title{tem_trans_finder}
\usage{
tem_trans_finder(o, f, c, t, s)
}
\arguments{
\item{o}{mean of observed temperature}

\item{f}{mean of future temperature}

\item{c}{mean of controlled scenario temperature}

\item{t}{observed temperature}

\item{s}{difference between standard deviation of future temperature and controlled temperature scenario you can find it by dif_finder function}
}
\value{

}
\description{
this function find the temperature transition
}
