% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem_var_plot_function.R
\name{tem_var_plot}
\alias{tem_var_plot}
\title{tem_var_plot}
\usage{
tem_var_plot(ss, sc)
}
\arguments{
\item{ss}{standard deviation of scenario period}

\item{sc}{standard deviation of control period}
}
\value{
a plot
}
\description{
it plots how the variability of tempereture changes
}
