% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twas.R
\name{twas}
\alias{twas}
\title{TWAS Builder}
\usage{
twas(data, dap)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}

\item{dap}{a dap object}
}
\value{
\code{twas} returns an object of \code{"twas"}, which is a list containing the following components:
\item{weight}{a data frame of variants and corresponding weights.}
\item{ER2}{the expected R-squared.}
}
\description{
\code{twas} calculates weights on variants.
}
\examples{
set.seed(0)
n = 100
p = 1000

a1 = rnorm(n)
a2 = 10*a1+rnorm(n)
a3 = 2*a1+9*a2+rnorm(n)

b1 = rnorm(n)
b2 = 8*b1+rnorm(n)

x = matrix(rnorm(n*(p-5)), nrow=n)

df = data.frame(a1,a2,a3,b1,b2,x)
df$y = 2*a1+b1+rnorm(n)

test.dap = dap(y~., df)
twas(df, test.dap)

}
