% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_log10_posterior}
\alias{compute_log10_posterior}
\title{Compute log10 posterior scores for multiple model configurations}
\usage{
compute_log10_posterior(X, y, cmfg_matrix, single_matrix, pi_vec, phi2_mat)
}
\arguments{
\item{X}{An n*p matrix of genotype data}

\item{y}{An n-vector of phenotype data}

\item{cmfg_matrix}{An m*p matrix of model configurations}

\item{pi_vec}{A vector of prior probabilities}

\item{phi2_vec}{A vector of phi2 values}
}
\value{
An m-vector of log10 posterior scores
}
\description{
Compute log10 posterior scores for multiple model configurations
}
