% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap.R
\name{dap}
\alias{dap}
\title{DAP fine mapping}
\usage{
dap(
  X,
  y,
  L = min(10, ncol(X)),
  prior_weights = NULL,
  null_weight = NULL,
  residual_tau = NULL,
  threshold = 1e-06,
  phi2 = 0.36,
  phi2_vec = NULL,
  r2_threshold = 0.25,
  coverage = NULL,
  option = 1
)
}
\arguments{
\item{X}{Genotype data}

\item{y}{Phenotype data}

\item{L}{Number of causal variants}

\item{prior_weights}{Prior weights}

\item{null_weight}{Null weight}

\item{residual_tau}{Residual variance}

\item{threshold}{Threshold for proposal density}

\item{phi2}{Scaled prior effect size variance}

\item{phi2_vec}{Scaled prior effect size variance vector}

\item{r2_threshold}{Genotype R2 threshold for LD}

\item{coverage}{Coverage of credible sets. When not set, it will output signal clusters; otherwise, it outputs credible sets at the specified coverage level}

\item{option}{Option for DAP. 1: default SuSiE; 2: SuSiE with residual variance estimation; 3: SuSiE with prior variance estimation}
}
\value{
Fine mapping results
}
\description{
DAP fine mapping
}
