% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap.R
\name{dap}
\alias{dap}
\title{DAP-PIR}
\usage{
dap(
  X,
  y,
  L = min(10, ncol(X)),
  prior_weights = NULL,
  null_weight = NULL,
  residual_tau = NULL,
  threshold = 1e-06,
  phi2 = 0.36,
  phi2_vec = NULL
)
}
\arguments{
\item{X}{Genotype data}

\item{y}{Phenotype data}

\item{L}{Number of causal variants}

\item{prior_weights}{Prior weights}

\item{null_weight}{Null weight}

\item{residual_tau}{Residual variance}

\item{threshold}{Threshold for proposal density}

\item{phi2}{Scaled prior effect size variance}

\item{phi2_vec}{Scaled prior effect size variance vector}
}
\value{
Fine mapping results
}
\description{
DAP-PIR
}
