% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dap.R
\name{dap}
\alias{dap}
\title{DAP-S fine-mapping with individual-level data.}
\usage{
dap(
  X,
  y,
  L = min(10, ncol(X)),
  susie_params = list(),
  prior_weights = NULL,
  standardize = TRUE,
  use_susie_variance_estimate = TRUE,
  grid = c(0.04, 0.16, 0.64),
  exclusive = TRUE,
  pir_threshold = 1e-06,
  r2_threshold = 0.25,
  coverage = NULL
)
}
\arguments{
\item{X}{Genotype matrix (n x p)}

\item{y}{Phenotype vector (n)}

\item{L}{Number of causal variants}

\item{susie_params}{Parameters for SuSiE: \code{scaled_prior_variance},
\code{residual_variance}, \code{null_weight}, \code{standardize},
\code{estimate_residual_variance}, \code{estimate_prior_variance}.}

\item{prior_weights}{Prior weights of each SNP being causal}

\item{standardize}{Standardize each column of the genotype matrix}

\item{use_susie_variance_estimate}{Use SuSiE-estimated prior variance}

\item{grid}{Grid of prior variances, default is c(0.04, 0.16, 0.64)}

\item{exclusive}{Exclusive signal clusters}

\item{pir_threshold}{Threshold for PIR}

\item{r2_threshold}{Threshold for R2 when constructing credible sets}

\item{coverage}{Coverage for credible set}

\item{scaled_prior_variance}{Scaled prior variance. That is, the prior
variance of each non-zero element of b is \code{var(y)} *
\code{scaled_prior_variance}.}

\item{residual_variance}{Residual variance}

\item{null_weight}{Null weight used in SuSiE}

\item{estimate_residual_variance}{Estimate residual variance in SuSiE}

\item{estimate_prior_variance}{Estimate prior variance in SuSiE}
}
\value{
List of fine-mapping results.
}
\description{
DAP-S fine-mapping with individual-level data.
}
