% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dap}
\alias{dap}
\title{Structured Bayesian Model Selection via Deterministic Approximation of Posteriors}
\usage{
dap(formula, data, ens=1, pi1=-1, ld_control=0.25, msize=-1, converg_thresh=0.01, all=FALSE, size_limit=-1, thread=1, quiet=FALSE)
}
\arguments{
\item{formula}{an object of class \code{formula} (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{a data frame containing the variables in the model.}

\item{ens}{(optional) prior expected number of signal clusters, \code{ens=1} by default.}

\item{pi1}{(optional) the exchangeable prior probability, values \code{0<pi1<1} accepted. By default -1, \code{pi1=ens/p}, where \code{p} is number of predictors in the input file.}

\item{ld_control}{(optional) the LD threshold to be considered within a single signal cluster. By default, the threshold is set to 0.25.}

\item{msize}{(optional) the maximum size of model dap explores. Valid maximum model size ranges from \code{1} to \code{p}. By default -1, it is set to \code{p}, i.e., there is no restriction on how large the true association model can be. If it is specified, the DAP runs DAP-K algorithm and stops at the specified maximum model size.}

\item{converg_thresh}{(optional)  the stopping condition for model exploration. By default, \code{converg_thresh=0.01}.}

\item{all}{(optional) If TRUE, dap will output information for all predictors and all signal clusters. By default, only predictors with \code{PIP > 0.001} and signal clusters with \code{SPIP > 0.25} are output.}

\item{thread}{(optional) the number of parallel threads to run DAP algorithm, 1 by default. OpenMP is required for multi-thread option.}

\item{quiet}{(optional) If TRUE, dap will mute running logs.}
}
\value{
\code{dap} returns an object of \code{"dap"}, which is a list containing the following components: \item{cluster}{a data frame, with each line representing the information of one signal cluster, including the size (i.e. number of member predictors), the posterior inclusion probability, and the average LD measures (\eqn{r^2}) for predictors within the cluster.}
\item{cluster.r2}{a matrix representing the average LD measures (\eqn{r^2}) for predictors within a cluster and between clusters.}
\item{signal}{a data frame of predictors ordered by the posterior inclusion probability (PIP), including predictor name, PIP, log10abf, and the signal cluster it belongs to.}
\item{model}{a data frame of models. Specifically, the first column shows the posterior probability of the corresponding model; the second column indicates the size (i.e., the number of predictors) of the model; the third column shows the unnormalized posterior score of the model (defined as \eqn{log10(model prior)+log10(BF)}); and the last column gives the exact configuration of the model.}
\item{info}{a list of extra information on the expected model size, sample size and the minimum PIP.}
\item{call}{the matched call.}
}
\description{
\code{dap} is used perform Bayesian variable selection among a large scale of predictors, multicolinearty and missingness are allowed. It will automatically impute missingness in the input data with mean values, normalize the response and predictors, and propose top predictive signals and signal clusters according to posterior probability.
}
\details{
Please refer to \url{https://github.com/xqwen/dap/tree/master/dap_src} for more details.
}
\examples{
set.seed(0)
n = 100
p = 1000

a1 = rnorm(n)
a2 = 10*a1+rnorm(n)
a3 = 2*a1+9*a2+rnorm(n)

b1 = rnorm(n)
b2 = 8*b1+rnorm(n)

x = matrix(rnorm(n*(p-5)), nrow=n)

df = data.frame(a1,a2,a3,b1,b2,x)
df$y = 2*a1+b1+rnorm(n)

test.dap = dap(y~., df)
test.dap

summary(test.dap)

}
\references{
Wen, X., Lee, Y., Luca, F., Pique-Regi, R. Efficient Integrative Multi-SNP Association Analysis using Deterministic Approximation of Posteriors. \emph{The American Journal of Human Genetics}, 98(6), 1114--1129

Lee, Y, Luca, F, Pique-Regi, R,Wen, X. Bayesian Multi-SNP Genetic Association Analysis: Control of FDR and Use of Summary Statistics bioRxiv:316471
}
\seealso{
\code{\link{summary.dap}} for summaries; and \code{\link{dap.sbams}} for a different interface directly analyzing a sbams-format file.
}
