% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dap_main}
\alias{dap_main}
\title{Implementation of DAP-S algorithm in C++}
\usage{
dap_main(
  matrix,
  pir_threshold,
  prior_weights,
  phi2_mat,
  r2_threshold,
  coverage,
  overlapping,
  twas_weight,
  snp_names,
  ss = 0L,
  X_input = NULL,
  y_input = NULL,
  XtX_input = NULL,
  Xty_input = NULL,
  yty_input = NULL,
  n_input = NULL
)
}
\arguments{
\item{matrix}{Proposal density matrix from SuSiE}

\item{pir_threshold}{Threshold for PIR}

\item{prior_weights}{Vector of prior probabilities}

\item{phi2_mat}{Matrix of scaled prior effect size variances}

\item{r2_threshold}{Threshold for LD}

\item{coverage}{Coverage for credible set}

\item{overlapping}{If TRUE. enforce overlapping clusters}

\item{twas_weight}{If TRUE, return TWAS weights}

\item{snp_names}{SNP names}

\item{X}{Genotype matrix}

\item{y}{Phenotype vector}
}
\value{
A list containing:
\itemize{
  \item model_config - Model configurations
  \item posterior_prob - Posterior probabilities
  \item log10_posterior_score - Log10 posterior scores
  \item log10_nc - Log10 normalizing constant
  \item pip - Posterior inclusion probabilities
  \item signal_cluster - Signal clusters
}
}
\description{
Implementation of DAP-S algorithm in C++
}
