% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbams.R
\name{read.sbams}
\alias{read.sbams}
\title{Read a sbams file}
\usage{
read.sbams(file)
}
\arguments{
\item{file}{file path to the sbams file}
}
\value{
a sbams object with pheno and geno, and controlled covariates if applicable.
}
\description{
This function will import the sbams file, automatically impute missing values with mean, and regress pheno and geno out of controlled covariants if applicable.
}
