% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbams.R
\name{read.sbams}
\alias{read.sbams}
\title{Read a sbams file}
\usage{
read.sbams(file)
}
\arguments{
\item{file}{file path to the sbams file}
}
\value{
a data.frame with the first colunm as the normalized phenotype, and the following columns as the normalized genotypes.
}
\description{
This function will import the sbams file, automatically impute missing values with mean, regress pheno and geno out of controlled covariants if applicable, and generate an \R data frame which can be passed into the \code{\link{dap}} function.
}
\details{
Please refer to \url{https://github.com/xqwen/dap/tree/master/dap_src} for more details.
}
\examples{
\dontrun{

# test data from https://github.com/xqwen/dap/blob/master/dap_src/sample_data/sim.1.sbams.dat
sbams.dat = read.sbams("sim.1.sbams.dat")
}
}
